+++ src/app/admin/edit/[tripId]/page.tsx
@@ -10,6 +10,7 @@ import { combineAccommodationDescriptions } from '@/app/lib/combineAccommodation
 import { formatLocalDateInput, getLocalDateSortValue } from '@/app/lib/localDateUtils';
 import DeleteWarningDialog from '@/app/admin/components/DeleteWarningDialog';
 import ReassignmentDialog from '@/app/admin/components/ReassignmentDialog';
+import ToastNotification from '@/app/admin/components/ToastNotification';
 import TripMetadataForm from './components/TripMetadataForm';
 import LocationManager from './components/LocationManager';
 import RouteManager from './components/RouteManager';
@@ -29,67 +30,6 @@ const EDITOR_SECTIONS: EditorSectionId[] = ['trip', 'locations', 'routes', 'acco
 
 const isEditorSectionId = (value: string | null): value is EditorSectionId =>
   value !== null && EDITOR_SECTIONS.includes(value as EditorSectionId);
-
-type ToastNotificationProps = {
-  notification: { message: string; type: 'success' | 'error' | 'info'; isVisible: boolean };
-  onClose: () => void;
-};
-
-function ToastNotification({ notification, onClose }: ToastNotificationProps) {
-  return (
-    <div className={`fixed top-4 right-4 z-50 transition-all duration-300 transform ${notification.isVisible ? 'translate-x-0 opacity-100' : 'translate-x-full opacity-0'
-      }`}>
-      <div className={`max-w-sm w-full shadow-lg rounded-lg pointer-events-auto ring-1 ring-black ring-opacity-5 overflow-hidden ${notification.type === 'success' ? 'bg-green-50 dark:bg-green-900' :
-        notification.type === 'error' ? 'bg-red-50 dark:bg-red-900' :
-          'bg-blue-50 dark:bg-blue-900'
-        }`}>
-        <div className="p-4">
-          <div className="flex items-start">
-            <div className="flex-shrink-0">
-              {notification.type === 'success' && (
-                <svg className="h-6 w-6 text-green-400" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z" />
-                </svg>
-              )}
-              {notification.type === 'error' && (
-                <svg className="h-6 w-6 text-red-400" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M10 14l2-2m0 0l2-2m-2 2l-2-2m2 2l2 2m7-2a9 9 0 11-18 0 9 9 0 0118 0z" />
-                </svg>
-              )}
-              {notification.type === 'info' && (
-                <svg className="h-6 w-6 text-blue-400" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M13 16h-1v-4h-1m1-4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
-                </svg>
-              )}
-            </div>
-            <div className="ml-3 w-0 flex-1 pt-0.5">
-              <p className={`text-sm font-medium ${notification.type === 'success' ? 'text-green-800 dark:text-green-200' :
-                notification.type === 'error' ? 'text-red-800 dark:text-red-200' :
-                  'text-blue-800 dark:text-blue-200'
-                }`}>
-                {notification.message}
-              </p>
-            </div>
-            <div className="ml-4 flex-shrink-0 flex">
-              <button
-                onClick={onClose}
-                className={`rounded-md inline-flex focus:outline-none focus:ring-2 focus:ring-offset-2 ${notification.type === 'success' ? 'text-green-500 hover:text-green-600 focus:ring-green-500' :
-                  notification.type === 'error' ? 'text-red-500 hover:text-red-600 focus:ring-red-500' :
-                    'text-blue-500 hover:text-blue-600 focus:ring-blue-500'
-                  }`}
-              >
-                <span className="sr-only">Close</span>
-                <svg className="h-5 w-5" viewBox="0 0 20 20" fill="currentColor">
-                  <path fillRule="evenodd" d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z" clipRule="evenodd" />
-                </svg>
-              </button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
 
 /**
  * Render the trip editor page for creating or editing a travel map, including metadata, locations, routes, accommodations, export, and admin access checks.
@@ -844,7 +784,7 @@ export default function TripEditorPage() {
             />
           )}
 
-          {/* Toast Notification - Moved from TravelDataForm.tsx */}
+          {/* Toast Notification */}
           {notification && (
             <ToastNotification
               notification={notification}
